third_party/WebKit/LayoutTests/typedcssom/cssVariableReferenceValue.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(function() {
    var expected = 'anything';
    var variableReferenceValue = new CSSVariableReferenceValue(expected);
    assert_true(variableReferenceValue instanceof CSSVariableReferenceValue);
    assert_equals(variableReferenceValue.variable, expected);
}, "The 'variable' attribute is correct");

</script>
